+++ src/components/RadioGroup/RadioGroup.test.tsx
@@ -309,19 +309,6 @@ describe("Radio Button", () => {
         </RadioGroup>
       )
       .toJSON();
-    const noOptReqLabel = renderer
-      .create(
-        <RadioGroup
-          labelText="no optional or required label"
-          name="optReq"
-          id="optReq"
-          optReqFlag={false}
-        >
-          <Radio value="2" labelText="Radio 2" id="radio-2" />
-          <Radio value="3" labelText="Radio 3" id="radio-3" />
-        </RadioGroup>
-      )
-      .toJSON();
     const isRequired = renderer
       .create(
         <RadioGroup
@@ -329,6 +316,20 @@ describe("Radio Button", () => {
           name="required"
           id="required"
           isRequired
+        >
+          <Radio value="2" labelText="Radio 2" id="radio-2" />
+          <Radio value="3" labelText="Radio 3" id="radio-3" />
+        </RadioGroup>
+      )
+      .toJSON();
+    const noRequiredLabel = renderer
+      .create(
+        <RadioGroup
+          labelText="no optional or required label"
+          name="noRequiredLabel"
+          id="noRequiredLabel"
+          isRequired
+          showRequiredLabel={false}
         >
           <Radio value="2" labelText="Radio 2" id="radio-2" />
           <Radio value="3" labelText="Radio 3" id="radio-3" />
@@ -395,7 +396,7 @@ describe("Radio Button", () => {
     expect(noLabel).toMatchSnapshot();
     expect(helperText).toMatchSnapshot();
     expect(invalidText).toMatchSnapshot();
-    expect(noOptReqLabel).toMatchSnapshot();
+    expect(noRequiredLabel).toMatchSnapshot();
     expect(isRequired).toMatchSnapshot();
     expect(isInvalid).toMatchSnapshot();
     expect(isDisabled).toMatchSnapshot();